[FIX] argument parsing for cutoff

+++ include/raptor/prepare/cutoff.hpp
@@ -36,7 +36,7 @@ public:
         }
     }
 
-    size_t get(std::filesystem::path const & filename) const noexcept
+    uint8_t get(std::filesystem::path const & filename) const noexcept
     {
         switch (cutoff_kind)
         {
@@ -65,20 +65,20 @@ private:
 
     cutoff_kinds cutoff_kind{cutoff_kinds::fixed};
 
-    uint16_t fixed_cutoff{};
+    uint8_t fixed_cutoff{};
     // Cutoffs and bounds from Mantis
     // Mantis ignores k-mers which appear less than a certain cutoff. The cutoff is based on the file size of a
     // gzipped fastq file. Small files have only a cutoff of 1 while big files have a cutoff value of 50.
     // https://doi.org/10.1016/j.cels.2018.05.021
     // Supplement Table S1
     // https://www.cell.com/cms/10.1016/j.cels.2018.05.021/attachment/0a3d402b-8b90-42c0-a709-22f246fd1759/mmc1.pdf
-    static constexpr std::array<uint16_t, 4> const cutoffs{1u, 3u, 10u, 20u};
+    static constexpr std::array<uint8_t, 4> const cutoffs{1u, 3u, 10u, 20u};
     static constexpr std::array<uint64_t, 4> const cutoff_bounds{314'572'800ULL,
                                                                  524'288'000ULL,
                                                                  1'073'741'824ULL,
                                                                  3'221'225'472ULL};
 
-    uint16_t impl(std::filesystem::path const & filename) const
+    uint8_t impl(std::filesystem::path const & filename) const
     {
         bool const is_compressed = file_is_compressed(filename);
         bool const is_fasta = check_for_fasta_format(filename);
@@ -88,7 +88,7 @@ private:
         // We divide by 3 if the input is not compressed.
         size_t const filesize = std::filesystem::file_size(filename) * (is_fasta ? 2 : 1) / (is_compressed ? 1 : 3);
 
-        uint16_t cutoff{50u};
+        uint8_t cutoff{50u};
         for (size_t i = 0; i < cutoff_bounds.size(); ++i)
         {
             if (filesize <= cutoff_bounds[i])
